clear roll
